test: fold the binding cases into the existing input-binding spec
A separate ui-input-binding.spec.tsx sat one character away from the
input-binding.spec.tsx that already covered this prop on UiEntity, which
is a good way to have the two edited out of step later.
